test: fold export compatibility tests into canonical test files
