+++ .superpowers/sdd/progress.md
@@ -1,0 +1,3 @@
+# Export Device Policy SDD Progress
+
+Task 1: complete (commits 4cecdf5c..f991880f, review clean; minor notes: policy reason wording, unused alias, catalog-coupled test).
